Adds git member to workspace
For now it contains the function `cat_file`
