+++ git/Cargo.toml
@@ -1,0 +1,8 @@
+[package]
+name = "git"
+version = "0.1.0"
+authors = ["Intent HQ <engineering@intenthq.com>"]
+edition = "2018"
+
+[dependencies]
+git2 = "0.8.*"
